+++ usr.sbin/amd/fsinfo/fsi_data.h
@@ -36,7 +36,7 @@
  * SUCH DAMAGE.
  *
  *	from: @(#)fsi_data.h	8.1 (Berkeley) 6/6/93
- *	$Id: fsi_data.h,v 1.1.1.1 1995/10/18 08:47:18 deraadt Exp $
+ *	$Id: fsi_data.h,v 1.2 1997/01/31 14:42:10 graichen Exp $
  */
 
 typedef struct auto_tree auto_tree;
@@ -120,7 +120,7 @@ struct ether_if {
 #define	EF_INADDR 0
 	struct in_addr e_inaddr;
 #define	EF_NETMASK 1
-	u_long e_netmask;
+	u_int32_t e_netmask;
 #define	EF_HWADDR 2
 	char *e_hwaddr;
 };
